+++ CHANGELOG.md
@@ -5,10 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.5]
+
+- upgraded to `khao-malet 2.0.3`
+
+## [3.0.4]
+
+- upgraded to `khao-malet 2.0.2`
+
+## [3.0.3]
+
+- upgraded to `khao-malet 2.0.1`
+
+## [3.0.2]
+
+- added `surface-high`, `surface-highest` , `surface-low`, `surface-lowest` as filling option to `khao-card`
+- set `surface` as new defult filling for `khao-card`
+
+## [3.0.1]
+
+- maded `khao-chip` type `outline` transparent
+
 ## [3.0.0]
 
-- use khao-malet 2.0.0 with new color schema and system
-- adjusted
+- use `khao-malet 2.0.0` with new color schema and system
+- adjusted colors
 
 ## [2.0.2]